+++ GRDB/DatabaseValue.swift
@@ -45,6 +45,9 @@ public enum DatabaseValue : Equatable {
     /// The BLOB storage class, wrapping a Blob.
     case Blob(GRDB.Blob)
     
+    
+    // MARK: - DatabaseValueConvertible Values
+    
     /**
     Returns the wrapped value.
     
@@ -121,5 +124,22 @@ public func ==(lhs: DatabaseValue, rhs: DatabaseValue) -> Bool {
         return lhs.data.isEqualToData(rhs.data)
     default:
         return false
+    }
+}
+
+/// A Database Blob
+public struct Blob {
+    
+    /// The data. Its length is guaranteed to be greater than zero.
+    public let data: NSData
+    
+    /// Creates a Blob from NSData. Returns nil if and only if *data* is nil or
+    /// zero-length (SQLite can't store empty blobs).
+    public init?(_ data: NSData?) {
+        if let data = data where data.length > 0 {
+            self.data = data
+        } else {
+            return nil
+        }
     }
 }