+++ GRDB/Statement.swift
@@ -144,10 +144,12 @@ public class Statement {
 // MARK: - SQLite identifier quoting
 
 extension String {
-    /// Returns the receiver, quoted for safe insertion in an SQL query as an
-    /// identifier.
-    ///
-    ///     db.execute("SELECT * FROM \(tableName.quotedDatabaseIdentifier)")
+    /**
+    Returns the receiver, quoted for safe insertion as an identifier in an SQL
+    query.
+
+        db.execute("SELECT * FROM \(tableName.quotedDatabaseIdentifier)")
+    */
     public var quotedDatabaseIdentifier: String {
         // See https://www.sqlite.org/lang_keywords.html
         return "\"\(self)\""